removed downloads folder from remote

+++ server/src/main/java/net/guides/springboot2/crud/services/RealizariRetineriService.java
@@ -171,7 +171,7 @@ public class RealizariRetineriService {
         rr.setValcm(valCM);
         rr.setZileplatite(zilePlatite);
         rr.setNroresuplimentare(nrOreSuplimentare);
-        rr.setSalariurealizat(salariuRealizat);
+        rr.setSalariurealizat((int)salariuRealizat);
 
         return rr;
     } // calcRealizariRetineri
